Better design for feedback
