+++ frontend/tamuhack/src/machineLearing.js
@@ -1,11 +1,13 @@
 import Chatbot from "./Bot.js";
 import Navbar from "./navbar.js";
+import FeedbackComponent from "./feedbackComponent.js";
 
 function MachineLearning() {
     return (
         <>
             <Navbar />
             <Chatbot />
+            <FeedbackComponent />
         </>
     )
 }
