+++ frontend/tamuhack/src/runways.js
@@ -1,5 +1,6 @@
 import Navbar from "./navbar.js";
 import Runway from "./runway.js";
+import FeedbackComponent from "./feedbackComponent.js";
 import { useEffect, useState } from "react";
 // import useFetch from "./useFetch.js";
 function Runways() {
@@ -46,6 +47,7 @@ function Runways() {
                     {renderRunways(runwayAvailability, flightNumbers)}
                 </tbody>
             </table>
+            <FeedbackComponent />
         </>
     )
 }
